Added new project box for Nike Shoes site; updated HTML, CSS, and JS to include new modal and image; updated arrays to include new elements.
This commit adds a new project box for the Nike Shoes site, which includes an image and a link to the site. The HTML, CSS, and JS files have been updated to include the new modal and image, as well as updated arrays to include the new elements. The new project box is hidden by default and can be displayed when the corresponding link is clicked. The close button on the new modal also works as expected.